part6/unicafe-redux/src/reducer.test.ts
import deepFreeze from 'deep-freeze'
import counterReducer from './reducer'
import { FeedbackActionType } from './types';

describe('unicafe reducer', () => {
  const initialState = {
    good: 0,
    ok: 0,
    bad: 0
  }

  test('should return a proper initial state when called with undefined state', () => {
    const state = {}
    const action = {
      type: FeedbackActionType.DO_NOTHING,
    }

    const newState = counterReducer(undefined, action)
    expect(newState).toEqual(initialState)
  })

  test('good is incremented', () => {
    const action = {
      type: FeedbackActionType.GOOD,
    }
    const state = initialState

    deepFreeze(state)
    const newState = counterReducer(state, action)
    expect(newState).toEqual({
      good: 1,
      ok: 0,
      bad: 0
    })
  })

  test('bad is incremented', () => {
    const action = {
      type: FeedbackActionType.BAD,
    };
    const state = initialState;

    deepFreeze(state);
    const newState = counterReducer(state, action);
    expect(newState).toEqual({
      good: 0,
      ok: 0,
      bad: 1,
    });
  });

  test('ok is incremented', () => {
    const action = {
      type: FeedbackActionType.OK,
    };
    const state = initialState;

    deepFreeze(state);
    const newState = counterReducer(state, action);
    expect(newState).toEqual({
      good: 0,
      ok: 1,
      bad: 0,
    });
  })

  test('zero resets the state', () => {
    const action = {
      type: FeedbackActionType.ZERO,
    };
    const state = {
      good: 1,
      ok: 2,
      bad: 3,
    };

    deepFreeze(state);
    const newState = counterReducer(state, action);
    expect(newState).toEqual(initialState);
  })
})